Remove gen_decl
