--- tutorlolv2_dev/src/generators/gen_decl/mod.rs
@@ -1,1 +1,0 @@
-pub mod decl_items;
