+++ tutorlolv2_dev/src/generators/gen_items/mod.rs
@@ -1,10 +1,7 @@
 pub(self) use crate::{
     MayFail, Progress,
     Progress::*,
-    generators::{
-        Generator, gen_decl::decl_items::*, gen_factories::fac_items::ItemData,
-        gen_utils::RegExtractor,
-    },
+    generators::{Generator, gen_factories::fac_items::ItemData, gen_utils::RegExtractor},
 };
 pub(self) use tutorlolv2_gen::{
     ItemId,
@@ -12,9 +9,788 @@ pub(self) use tutorlolv2_gen::{
     eval::CtxVar::*,
 };
 
-pub mod blade_of_the_ruined_king;
-pub mod blade_of_the_ruined_king_arena;
-pub mod malignance;
-pub mod malignance_arena;
-pub mod nashors_tooth;
-pub mod nashors_tooth_arena;
+macro_rules! decl_items {
+    (inner $Name:ident) => {
+        pub struct $Name(pub ItemData);
+
+        impl $Name {
+            pub fn new(data: ItemData) -> Box<dyn Generator<ItemData>> {
+                Box::new(Self(data))
+            }
+
+            pub fn end(self) -> MayFail<ItemData> {
+                Ok(self.0)
+            }
+
+            pub fn into_inner(self) -> ItemData {
+                self.0
+            }
+        }
+
+        impl ::core::ops::Deref for $Name {
+            type Target = ItemData;
+            fn deref(&self) -> &Self::Target {
+                &self.0
+            }
+        }
+
+        impl ::core::ops::DerefMut for $Name {
+            fn deref_mut(&mut self) -> &mut Self::Target {
+                &mut self.0
+            }
+        }
+    };
+    (
+        impl {
+            $($Stable:ident),*$(,)*
+        },
+        default {
+            $($Name:ident),*$(,)*
+        }
+    ) => {
+        $(
+            decl_items!(inner $Name);
+
+            impl Generator<ItemData> for $Name {
+                fn generate(mut self: Box<Self>) -> MayFail<ItemData> {
+                    self.infer_stats_ifdef();
+                    self.end()
+                }
+            }
+        )*
+        $(
+            pastey::paste! {
+                pub mod [<$Stable:snake>];
+            }
+
+            decl_items!(inner $Stable);
+        )*
+
+        pub fn item_gen_fn(item_id: &str) -> Option<
+            fn(ItemData) -> Box<dyn Generator<ItemData>>
+        > {
+            match item_id {
+                $(stringify!($Name) => Some($Name::new),)*
+                $(stringify!($Stable) => Some($Stable::new),)*
+                _ => None,
+            }
+        }
+
+        pub const fn item_gen_names() -> &'static [&'static str] {
+            &[
+                $(stringify!($Name),)*
+                $(stringify!($Stable),)*
+            ]
+        }
+
+        pub const fn item_gen_riot_ids() -> [u32; item_gen_names().len()] {
+            [
+                $(ItemId::$Name.to_riot_id(),)*
+                $(ItemId::$Stable.to_riot_id(),)*
+            ]
+        }
+    };
+}
+
+decl_items!(
+    impl {
+        BladeOfTheRuinedKing,
+        BladeOfTheRuinedKingArena,
+        Malignance,
+        MalignanceArena,
+        NashorsTooth,
+        NashorsToothArena
+    },
+    default {
+        AbyssalMask,
+        AbyssalMaskArena,
+        AbyssalMaskU32,
+        Actualizer,
+        ActualizerArena,
+        AdaptiveHelmArena,
+        AegisOfTheLegion,
+        AegisOfTheLegionArena,
+        AetherWisp,
+        AmplifyingTome,
+        AnathemasChains,
+        AnathemasChainsArena,
+        AnimaEcho,
+        Animapocalypse,
+        AniMines,
+        AntiSharkSeaMine,
+        AntiTowerSocks,
+        AnvilVoucherArena,
+        ArcaneSweeper,
+        ArchangelsStaff,
+        ArchangelsStaffArena,
+        ArchangelsStaffU32,
+        ArdentCenser,
+        ArdentCenserArena,
+        ArdentCenserU32,
+        ArmoredAdvance,
+        AtmasReckoning,
+        AtmasReckoningArena,
+        AtmasReckoningU66,
+        AxiomArc,
+        AxiomArcArena,
+        BamisCinder,
+        BandleglassMirror,
+        BandleJuiceOfHaste,
+        BandleJuiceOfPower,
+        BandleJuiceOfVitality,
+        Bandlepipes,
+        BandlepipesArena,
+        BansheesVeil,
+        BansheesVeilArena,
+        Bastionbreaker,
+        BattleBunnyCrossbow,
+        BattleCatBarrage,
+        BearfootChemDispenser,
+        BerserkersGreaves,
+        BerserkersGreavesArena,
+        BFSword,
+        BFSwordArena,
+        BlackCleaver,
+        BlackCleaverArena,
+        BlackfireTorch,
+        BlackfireTorchArena,
+        BlackHoleGauntletU44,
+        BlackMistScythe,
+        BladeORang,
+        BlastingWand,
+        BlastingWandArena,
+        BlightingJewel,
+        BloodlettersCurse4010,
+        BloodlettersCurse8010,
+        Bloodsong,
+        Bloodthirster,
+        BloodthirsterArena,
+        Boots,
+        BootsOfSwiftness,
+        BootsOfSwiftnessArena,
+        BotLaneQuest1202,
+        BotLaneQuest1207,
+        BountyOfWorlds,
+        BrambleVest,
+        BraveryVoucherArena,
+        BulwarkOfTheMountain,
+        BunnyHop,
+        BunnyMegaBlast,
+        BunnyPrimeBallista,
+        CappaJuice,
+        CappaJuiceArena,
+        CarrotCrash,
+        CatalystOfAeons,
+        CaulfieldsWarhammer,
+        CelestialOpposition,
+        ChainlacedCrushers,
+        ChainVest,
+        ChainVestArena,
+        ChaliceOfBlessing,
+        ChempunkChainsword,
+        ChempunkChainswordArena,
+        ChemtechPutrifier,
+        ChemtechPutrifierArena,
+        CloakOfAgility,
+        CloakOfStarryNightU44,
+        CloakOfStarryNightU66,
+        ClothArmor,
+        ControlWard,
+        CorruptingPotion,
+        CosmicDrive,
+        CosmicDriveArena,
+        CrimsonLucidity,
+        CrownOfTheShatteredQueen,
+        CrownOfTheShatteredQueenArena,
+        CrownOfTheShatteredQueenU44,
+        CrownOfTheShatteredQueenU66,
+        CrueltyU44,
+        CrueltyU66,
+        Cryptbloom,
+        CryptbloomArena,
+        CrystallineBracer,
+        Cull,
+        CyclonicSlicers,
+        Dagger,
+        DarkSeal,
+        DarksteelTalonsU44,
+        Dawncore,
+        DawncoreArena,
+        DawncoreU32,
+        DeadMansPlate,
+        DeadMansPlateArena,
+        DeathbladeArena,
+        DeathfireGrasp,
+        DeathsDance,
+        DeathsDanceArena,
+        DecapitatorU44,
+        DeepFreeze,
+        DemonicEmbrace,
+        DemonicEmbraceArena,
+        DemonicEmbraceU44,
+        DemonKingsCrownU44,
+        DemonKingsCrownU66,
+        DetonationOrbU44,
+        DiademOfSongs,
+        DiademOfSongsArena,
+        DiademOfSongsU32,
+        DiamondTippedSpearU44,
+        DisabledRecall,
+        DivineSunderer,
+        DivineSundererArena,
+        DivineSundererU44,
+        DoransBlade,
+        DoransRing,
+        DoransShield,
+        DoubleBunBunBarrage,
+        DragonheartU44,
+        DreamMaker,
+        DuskAndDawn,
+        DuskAndDawnArena,
+        DuskbladeOfDraktharr,
+        DuskbladeOfDraktharrArena,
+        DuskbladeOfDraktharrU44,
+        EchoesOfHelia,
+        EchoesOfHeliaArena,
+        EchoesOfHeliaU32,
+        EchoingBatblades,
+        Eclipse,
+        EclipseArena,
+        EdgeOfNight,
+        EdgeOfNightArena,
+        EleisasMiracleU44,
+        ElixirOfAvarice,
+        ElixirOfForce,
+        ElixirOfIron,
+        ElixirOfSkill,
+        ElixirOfSorcery,
+        ElixirOfWrath,
+        Emberknife,
+        EmpyreanPromiseU44,
+        EndlessHunger,
+        EndlessHungerArena,
+        EnhancedLuckyDice,
+        EnhancedRecall,
+        EnvelopingLight,
+        EssenceReaver,
+        EssenceReaverArena,
+        Evenshroud,
+        EvenshroudArena,
+        Everfrost,
+        EverfrostArena,
+        EverfrostU44,
+        EvolvedEmbershot,
+        ExecutionersCalling,
+        ExperimentalHexplate,
+        ExperimentalHexplateArena,
+        ExplosiveEmbrace,
+        EyeOfTheHerald1104,
+        EyeOfTheHerald3513,
+        FaerieCharm,
+        FarsightAlteration,
+        FatedAshes,
+        FcLimitedExpress,
+        FiendhunterBolts,
+        FiendhunterBoltsArena,
+        FiendishCodex,
+        Fimbulwinter,
+        FimbulwinterArena,
+        FimbulwinterU32,
+        FinalCityTransit,
+        FlesheaterU44,
+        FlesheaterU66,
+        ForbiddenIdol,
+        ForceOfEntropyU44,
+        ForceOfNature,
+        ForceOfNatureArena,
+        ForeverForward,
+        Fortification1501,
+        Fortification1521,
+        Frostfang,
+        FrozenHeart,
+        FrozenHeartArena,
+        FrozenHeartU32,
+        FrozenMallet,
+        FulminationU44,
+        Galeforce,
+        GaleforceArena,
+        GaleforceU44,
+        GamblersBladeU44,
+        GamblersBladeU66,
+        GangplankPlaceholder,
+        GargoyleStoneplate,
+        GargoyleStoneplateArena,
+        GargoyleStoneplateU44,
+        GargoyleStoneplateU66,
+        GatlingBunnyGuns,
+        Ghostcrawlers,
+        GhostcrawlersArena,
+        GiantsBelt,
+        GiantsBeltArena,
+        GlacialBuckler,
+        GlowingMote,
+        GlowingMoteArena,
+        GoldenSpatulaU99,
+        GoldStatAnvilVoucherArena,
+        Goredrinker,
+        GoredrinkerArena,
+        GrizzlySmash,
+        GuardianAngel,
+        GuardianAngelArena,
+        GuardiansAmulet,
+        GuardiansBlade,
+        GuardiansBladeArena,
+        GuardiansDirkArena,
+        GuardiansHammer,
+        GuardiansHammerArena,
+        GuardiansHorn,
+        GuardiansHornArena,
+        GuardiansOrb,
+        GuardiansOrbArena,
+        GuardiansShroud,
+        GuidingHex,
+        GuinsoosRageblade,
+        GuinsoosRagebladeArena,
+        GunmetalGreaves,
+        Gusto,
+        GustwalkerHatchling1102,
+        GustwalkerHatchling1106,
+        Hailblade,
+        HamstringerU44,
+        HarrowingCrescent,
+        HauntingGuise,
+        HealthPotion,
+        HearthboundAxe,
+        Heartsteel,
+        HeartsteelArena,
+        HellfireHatchet,
+        HemomancersHelmU44,
+        HexboltCompanionU44,
+        Hexdrinker,
+        HexopticsC44,
+        HexopticsC44Arena,
+        HextechAlternator,
+        HextechGunblade,
+        HextechGunbladeArena,
+        HextechGunbladeU66,
+        HextechRocketbelt,
+        HextechRocketbeltArena,
+        HollowRadiance,
+        HollowRadianceArena,
+        HoppedUpHex,
+        HorizonFocus,
+        HorizonFocusArena,
+        Hubris126697,
+        Hubris6697,
+        HubrisArena,
+        Hullbreaker,
+        HullbreakerArena,
+        IceblastArmor,
+        IcebornGauntlet,
+        IcebornGauntletArena,
+        ImmortalShieldbow,
+        ImmortalShieldbowArena,
+        ImperialMandate,
+        ImperialMandateArena,
+        ImperialMandateU32,
+        InfinityEdge,
+        InfinityEdgeArena,
+        InnervatingLocket,
+        InnervatingLocketU44,
+        IonianBootsOfLucidity,
+        IonianBootsOfLucidityArena,
+        IronspikeWhip,
+        JakShoTheProtean,
+        JakShoTheProteanArena,
+        JarvanIs,
+        JinxsTriNamite,
+        JuiceOfHaste,
+        JuiceOfPower,
+        JuiceOfVitality,
+        JungleQuest1204,
+        JungleQuest1209,
+        KaenicRookern,
+        KaenicRookernArena,
+        KalistasBlackSpear3599,
+        KalistasBlackSpear3600,
+        Kindlegem,
+        KindlegemArena,
+        KinkouJitteU44,
+        KircheisShard,
+        KnightsVow,
+        KnightsVowArena,
+        KnightsVowU32,
+        KrakenSlayer,
+        KrakenSlayerArena,
+        LastWhisper,
+        LeechingLeer,
+        LegendaryAssassinItemArena,
+        LegendaryFighterItemArena,
+        LegendaryMageItemArena,
+        LegendaryMarksmanItemArena,
+        LegendarySupportItemArena,
+        LegendaryTankItemArena,
+        LiandrysAnguishArena,
+        LiandrysTorment,
+        LichBane,
+        LichBaneArena,
+        Lifeline,
+        LifewellPendant,
+        LightningBraid,
+        LightningRodU44,
+        LightOfTheLion,
+        LionesssLament,
+        LockedWeaponSlot,
+        LocketOfTheIronSolari,
+        LocketOfTheIronSolariArena,
+        LocketOfTheIronSolariU32,
+        LongSword,
+        LordDominiksRegards,
+        LordDominiksRegardsArena,
+        LostChapter,
+        LoversRicochet,
+        LucentSingularity,
+        LuckyDice,
+        LudensEcho,
+        LudensEchoArena,
+        Manamune,
+        ManamuneArena,
+        ManamuneU32,
+        MawOfMalmortius,
+        MawOfMalmortiusArena,
+        MejaisSoulstealer,
+        MeowMeow,
+        MercurialScimitar,
+        MercurialScimitarArena,
+        MercurysTreads,
+        MercurysTreadsArena,
+        MidLaneQuest1201,
+        MidLaneQuest1206,
+        MikaelsBlessing,
+        MikaelsBlessingArena,
+        MikaelsBlessingU32,
+        MinionDematerializer,
+        MirageBladeU44,
+        MobilityBoots,
+        MoonflairSpellbladeU44,
+        MoonstoneRenewer,
+        MoonstoneRenewerArena,
+        MoonstoneRenewerU32,
+        Morellonomicon,
+        MorellonomiconArena,
+        MortalReminder,
+        MortalReminderArena,
+        MosstomperSeedling1103,
+        MosstomperSeedling1105,
+        MultitoolArena,
+        Muramana,
+        MuramanaArena,
+        MuramanaU32,
+        NavoriFlickerblade,
+        NavoriFlickerbladesArena,
+        NeedlesslyLargeRod,
+        NeedlesslyLargeRodArena,
+        NegatronCloak,
+        NegatronCloakArena,
+        NeverendingMobstomper,
+        NightHarvester,
+        NightHarvesterArena,
+        NightHarvesterU44,
+        Noonquiver,
+        NullMagicMantle,
+        OblivionOrb,
+        ObsidianCleaverArena,
+        ObsidianEdge,
+        Opportunity,
+        OpportunityArena,
+        OracleLens,
+        Overcharged1507,
+        Overcharged1523,
+        Overerchargedha,
+        Overgrowth,
+        OverlordsBloodmail,
+        OverlordsBloodmailU44,
+        OwoBlaster,
+        PartyFavor,
+        PauldronsOfWhiterock,
+        PawPrintPoisoner,
+        PenetratingBullets,
+        Perplexity,
+        Phage,
+        PhantomDancer,
+        PhantomDancerArena,
+        PhreakishGusto,
+        Pickaxe,
+        PillorySwipe,
+        PlatedSteelcaps,
+        PlatedSteelcapsArena,
+        PoroSnax,
+        PrismaticItemArena,
+        PrismaticStatVoucherArena,
+        ProfaneHydra,
+        ProfaneHydraArena,
+        ProtoplasmHarness,
+        ProtoplasmHarnessArena,
+        ProwlersClaw,
+        ProwlersClawArena,
+        ProwlersClawU44,
+        PrumbissElectrocarver,
+        PuppeteerU44,
+        PyromancersCloakU44,
+        QuadORang,
+        QuestBot,
+        QuestJungle,
+        QuestMid,
+        QuestSupport,
+        QuestTop,
+        QuicksilverSash,
+        RabadonsDeathcap,
+        RabadonsDeathcapArena,
+        RadiantField,
+        RadiantVirtue,
+        RadiantVirtueArena,
+        RadiantVirtueU44,
+        Rageknife,
+        RanduinsOmen,
+        RanduinsOmenArena,
+        RapidFirecannon,
+        RapidFirecannonArena,
+        RapidRabbitRaindown,
+        RavenousHydra,
+        RavenousHydraArena,
+        RealityFractureU44,
+        ReapersTollU44,
+        Recall,
+        Rectrix,
+        RecurveBow,
+        RecurveBowArena,
+        Redemption,
+        RedemptionArena,
+        RedemptionU32,
+        RefillablePotion,
+        RegicideU44,
+        ReinforcedArmor1502,
+        ReinforcedArmor1505,
+        ReinforcedArmor1506,
+        RejuvenationBead,
+        RelicShield,
+        ReverberationU44,
+        Riftmaker,
+        RiftmakerArena,
+        RiteOfRuin,
+        RodOfAges,
+        RodOfAgesArena,
+        RodOfAgesU32,
+        RubyCrystal,
+        RunaansHurricane,
+        RunaansHurricaneArena,
+        RunecarverU44,
+        RuneglaiveArena,
+        RunesteelSpaulders,
+        RunicCompass,
+        RylaisCrystalScepter,
+        RylaisCrystalScepterArena,
+        SanguineBladeArena,
+        SanguineGiftU44,
+        SapphireCrystal,
+        SavageSlice,
+        ScarecrowEffigy,
+        ScorchclawPup1101,
+        ScorchclawPup1107,
+        ScoutingAhead,
+        ScoutsSlingshot,
+        SearingShortbow,
+        SeekersArmguard,
+        SeraphsEmbrace,
+        SeraphsEmbraceArena,
+        SeraphsEmbraceU32,
+        SerpentsFang,
+        SerpentsFangArena,
+        SerratedDirk,
+        SeryldasGrudge,
+        SeryldasGrudgeArena,
+        Shadowflame,
+        ShadowflameArena,
+        ShardOfTrueIce,
+        ShatteredArmguard,
+        Sheen,
+        SheenArena,
+        ShieldOfMoltenStoneU44,
+        ShieldOfMoltenStoneU66,
+        ShieldOfTheRakkor,
+        ShieldSlam,
+        ShurelyasBattlesong,
+        ShurelyasBattlesongArena,
+        ShurelyasBattlesongU32,
+        SilvermereDawn,
+        SilvermereDawnArena,
+        SinEater,
+        SlightlyMagicalFootwear,
+        SmallPartyFavor,
+        SolsticeSleigh,
+        SorcerersShoes,
+        SorcerersShoesArena,
+        SoundWave,
+        SpearOfShojin,
+        SpearOfShojinArena,
+        SpectralCutlass,
+        SpectralCutlassArena,
+        SpectralSickle,
+        SpectresCowl,
+        SpellslingersShoes,
+        SpellthiefsEdge,
+        SpiritVisage,
+        SpiritVisageArena,
+        StaffOfFlowingWater,
+        StaffOfFlowingWaterArena,
+        StaffOfFlowingWaterU32,
+        StatBonus,
+        StatBonusArena,
+        StatikkShiv,
+        StatikkShivArena,
+        StatikkSword,
+        StealthWard2056,
+        StealthWard3340,
+        SteelShoulderguards,
+        SteelSigil,
+        SteelTempest,
+        SteraksGage,
+        SteraksGageArena,
+        StirringWardstone,
+        Stormrazor3095,
+        Stormrazor3097,
+        StormrazorArena,
+        Stormsurge,
+        StormsurgeArena,
+        Stridebreaker,
+        StridebreakerArena,
+        StructureBounty1516,
+        StructureBounty1517,
+        StructureBounty1518,
+        StructureBounty1519,
+        SunderedSky,
+        SunderedSkyArena,
+        SunfireAegis,
+        SunfireAegisArena,
+        SuperMechArmor,
+        SuperMechPowerField,
+        SupportQuest1203,
+        SupportQuest1208,
+        Swiftmarch,
+        SwordOfBlossomingDawn124011,
+        SwordOfBlossomingDawn4011,
+        SwordOfBlossomingDawnU66,
+        SwordOfTheDivine,
+        SwordOfTheDivineU44,
+        SwordOfTheDivineU66,
+        SymbioticSoles,
+        SynchronizedSouls,
+        TalismanOfAscensionU44,
+        TargonsBuckler,
+        TearOfTheGoddess,
+        TearOfTheGoddessU32,
+        TempestsGauntlet,
+        TentacleSlam,
+        Terminus,
+        TerminusArena,
+        TheAnnihilator,
+        TheBrutalizer,
+        TheCollector,
+        TheCollectorArena,
+        TheCollectorU66,
+        TheGoldenSpatula,
+        TheGoldenSpatulaArena,
+        TheGoldenSpatulaU66,
+        Thornmail,
+        ThornmailArena,
+        ThornmailU32,
+        Tiamat,
+        TIBBERS,
+        TIBBERSBEEGEdition,
+        TitanicHydra,
+        TitanicHydraArena,
+        TopLaneQuest1200,
+        TopLaneQuest1221,
+        TopLaneQuest1222,
+        TotalBiscuitOfEverlastingWill,
+        TowerPowerUp,
+        Trailblazer,
+        TrailblazerArena,
+        TrailblazerU32,
+        TrinityForce,
+        TrinityForceArena,
+        Tunneler,
+        TurboChemtankU44,
+        TurretPlating,
+        TwilightsEdgeU44,
+        TwinMaskU44,
+        UmbralGlaive,
+        UnceasingCyclone,
+        UnendingDespair,
+        UnendingDespairArena,
+        Unknown1205,
+        Unknown1210,
+        Unknown1211,
+        Unknown1212,
+        Unknown1213,
+        Unknown1214,
+        Unknown1215,
+        Unknown1216,
+        Unknown1217,
+        Unknown1218,
+        Unknown1219,
+        Unknown2008,
+        Unknown3901,
+        Unknown3902,
+        Unknown3903,
+        UnleashedTeleport,
+        UwuBlaster,
+        VampiricScepter,
+        VampiricScepterArena,
+        Vanguard,
+        VaynesChromablades,
+        VeigarsTalismanOfAscensionU66,
+        VerdantBarrier,
+        VigilantWardstone,
+        VoidStaff,
+        VoidStaffArena,
+        VoltaicCyclosword,
+        VoltaicCycloswordArena,
+        VortexGlove,
+        WanderingStorms,
+        WardensEye,
+        WardensMail,
+        WarmogsArmor,
+        WarmogsArmorU44,
+        WatchfulWardstone,
+        WhisperingCirclet,
+        WhisperingCircletArena,
+        WhisperingCircletU32,
+        WingedDagger,
+        WingedMoonplate,
+        WintersApproach,
+        WintersApproachArena,
+        WintersApproachU32,
+        WitsEnd,
+        WitsEndArena,
+        WoogletsWitchcapArena,
+        WordlessPromise,
+        WorldAtlas,
+        YoumuusGhostblade,
+        YoumuusGhostbladeArena,
+        YourCut,
+        YunTalWildarrows,
+        YunTalWildarrowsArena,
+        Yuumibot,
+        YuumibotFinalFinal,
+        ZazZaksRealmspike,
+        Zeal,
+        ZekesConvergence,
+        ZekesConvergenceArena,
+        ZekesConvergenceU32,
+        ZephyrArena,
+        ZephyrU66,
+        ZhonyasHourglass,
+        ZhonyasHourglassArena,
+    }
+);
